# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aasha Mental Health App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateCheckinRequest {
  sessionId: string;
  attendedClass: boolean;
  ateWell: boolean;
  /**
   * @minimum 1
   * @maximum 5
   */
  maskingLevel: number;
  /** How long user held the stone in milliseconds */
  holdDurationMs: number;
  /** Average milliseconds between micro-movements (cognitive load indicator) */
  interactionLatencyMs: number;
  lat?: number | null;
  lon?: number | null;
}

export interface Checkin {
  id: number;
  sessionId: string;
  attendedClass: boolean;
  ateWell: boolean;
  maskingLevel: number;
  holdDurationMs: number;
  interactionLatencyMs: number;
  isLateNight: boolean;
  lat?: number | null;
  lon?: number | null;
  createdAt: string;
}

export interface GardenData {
  totalPetals: number;
  recentCheckins: Checkin[];
  currentStreak: number;
}

export interface WeatherData {
  temperature: number;
  description: string;
  uvIndex: number;
  sunlightHours: number;
  barometricPressure: number;
  isLowSunlight: boolean;
  city: string;
}

export interface InsightRequest {
  sessionId: string;
  weatherData?: WeatherData;
  recentCheckins: Checkin[];
  academicWeek?: number | null;
}

export type InsightResponseCognitiveLoad =
  (typeof InsightResponseCognitiveLoad)[keyof typeof InsightResponseCognitiveLoad];

export const InsightResponseCognitiveLoad = {
  low: "low",
  moderate: "moderate",
  high: "high",
} as const;

export interface InsightResponse {
  note: string;
  cognitiveLoad: InsightResponseCognitiveLoad;
  showLightenLoad: boolean;
  sanctuarySuggestion?: string | null;
}

export interface ExtensionEmailRequest {
  professorName?: string | null;
  courseName?: string | null;
  assignmentName?: string | null;
  studentName?: string | null;
}

export interface ExtensionEmailResponse {
  subject: string;
  body: string;
  mailtoLink: string;
}

export interface CommunityPulse {
  totalUsers: number;
  inDarkStretch: number;
  percentageDarkStretch: number;
  averageMaskingLevel: number;
  message: string;
}

export type GetCheckinsParams = {
  sessionId: string;
  limit?: number;
};

export type GetGardenParams = {
  sessionId: string;
};

export type GetWeatherParams = {
  lat: number;
  lon: number;
};