# lib/api-zod/src/generated/types/checkin.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aasha Mental Health App API
 * OpenAPI spec version: 0.1.0
 */

export interface Checkin {
  id: number;
  sessionId: string;
  attendedClass: boolean;
  ateWell: boolean;
  maskingLevel: number;
  holdDurationMs: number;
  interactionLatencyMs: number;
  isLateNight: boolean;
  lat?: number | null;
  lon?: number | null;
  createdAt: Date;
}