# lib/api-zod/src/generated/types/communityPulse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aasha Mental Health App API
 * OpenAPI spec version: 0.1.0
 */

export interface CommunityPulse {
  totalUsers: number;
  inDarkStretch: number;
  percentageDarkStretch: number;
  averageMaskingLevel: number;
  message: string;
}