# lib/api-zod/src/generated/types/createCheckinRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aasha Mental Health App API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCheckinRequest {
  sessionId: string;
  attendedClass: boolean;
  ateWell: boolean;
  /**
   * @minimum 1
   * @maximum 5
   */
  maskingLevel: number;
  /** How long user held the stone in milliseconds */
  holdDurationMs: number;
  /** Average milliseconds between micro-movements (cognitive load indicator) */
  interactionLatencyMs: number;
  lat?: number | null;
  lon?: number | null;
}