# lib/api-zod/src/generated/types/extensionEmailRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aasha Mental Health App API
 * OpenAPI spec version: 0.1.0
 */

export interface ExtensionEmailRequest {
  professorName?: string | null;
  courseName?: string | null;
  assignmentName?: string | null;
  studentName?: string | null;
}