# lib/api-zod/src/generated/types/gardenData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aasha Mental Health App API
 * OpenAPI spec version: 0.1.0
 */
import type { Checkin } from "./checkin";

export interface GardenData {
  totalPetals: number;
  recentCheckins: Checkin[];
  currentStreak: number;
}