# lib/api-zod/src/generated/types/insightRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aasha Mental Health App API
 * OpenAPI spec version: 0.1.0
 */
import type { Checkin } from "./checkin";
import type { WeatherData } from "./weatherData";

export interface InsightRequest {
  sessionId: string;
  weatherData?: WeatherData;
  recentCheckins: Checkin[];
  academicWeek?: number | null;
}