# lib/api-zod/src/generated/types/insightResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aasha Mental Health App API
 * OpenAPI spec version: 0.1.0
 */
import type { InsightResponseCognitiveLoad } from "./insightResponseCognitiveLoad";

export interface InsightResponse {
  note: string;
  cognitiveLoad: InsightResponseCognitiveLoad;
  showLightenLoad: boolean;
  sanctuarySuggestion?: string | null;
}