# lib/api-zod/src/generated/types/weatherData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aasha Mental Health App API
 * OpenAPI spec version: 0.1.0
 */

export interface WeatherData {
  temperature: number;
  description: string;
  uvIndex: number;
  sunlightHours: number;
  barometricPressure: number;
  isLowSunlight: boolean;
  city: string;
}